Add a simple search API.
For now, need to go to /NAVADMIN and pass a query parameter 'q' set to
the search you want to perform (just a simple substring comparison).

A fun search was for 'housing' which lead to a NAVADMIN for the "A B C's
OF BASIC ALLOWANCE FOR HOUSING (BAH)" (NAVADMIN 083/08) which... man.