Fix subject detection splitting up across multiple NAVADMINs.
The Perl range operator ( /blah/ .. /blah/ ) we use to filter out
possible subject lines retains its configuration across new files. This
means that files that never trigger the terminator condition (any line
ending in '//', whitespace excluded) cause the test to continue to
return "true" for the beginning lines of the next file read.

This error manifests in NAVADMINs that are read after "PERSONAL FOR"
messages (which typically do not have a valid message terminator) to
have detected subject lines of the form "UNCLASSIFIED// ... SUBJ/ .." at
which point the error is reset.

To fix this I just do the easy thing: force the file contents to end in
a "//" even if they already do. This doesn't appear to break subject
detection and fixes the messages that come after PERSONAL FOR messages.