Fix download script for updated MyNavy HR website.
The new website, besides having a different base URL, appears to have
changed the format for some (not all) of the hyperlinks, which have a
query string like ?ver=<gibberish> appended.

On the other hand, the updated web server seems to respect the
If-Modified-Since header, so that's nice.

As a reminder the script is coded to *only* download the current year if
the file .has-run exists, and that file is generated on each run. This
is an attempt to be nice to the web server.

But it maybe prudent to re-run every so often. When the script pulled
down 2021 NAVADMINs again at least one had changed.