docker: Streamline the default image.
We use standard Perl modules that are already packaged so there's not
really any reason to use the `cpanm` client as we need to do with
perl-slim, which if I set it up to use system Perl would end up with two
Perls. Instead shift to alpine and its perl and perl module packages.