Add new and refresh existing NAVADMINs
I noticed I was missing NAVADMIN 315/18 so I decided to force re-running
my NAVADMIN downloader (which normally only updates NAVADMINs for the
current year out of respect for the upstream web server, under the
assumption that NAVADMINs don't change years later).

This also discovered a fair bit of other NAVADMINs that had been missed,
along with some minor changes to existing ones. Normally whitespace
changes but there was at least one where the date-time group of the
original message was apparently wrong by an entire month and fixed when
NPC did the web server upgrade.

This now appears correct as far back as the current website goes (by
checking the last issued NAVADMIN each year with the last issued
NAVADMIN in this list), *with the notable exception of the 2018
NAVADMINs*... although it was NAVADMIN 315/18 that got me looking, and
was fixed by running the downloader, NAVADMIN 316/18 through NAVADMIN
322/18 are still somehow missing...

But I'll investigate that for the next commit.