Fix the download script to look for NAVADMINs with lowercase URLs.
This helps to find the 2018 NAVADMINs following 315/18 (including some
in the middle of the range!). Sort of annoying that this was
inconsistent but to be honest I'm actually impressed that the vast
majority of URLs had been as consistent as they were.

I've verified that only the 2018 NAVADMINs in the total 1,695 or so
currently on the MyNavy HR website were affected.