Add a script to reformat broken RMKS headers.
The goal is to get the program to a point where it can read and
potentially convert NAVADMIN text from the appropriate sections into a
more approachable HTML version (including references to hyperlinks,
reformatting paragraphs with <ul>/<ol>, and so on).

But for now just see if we can find where the main text starts, which is
usually a RMKS/ header.

I use a script to do easy fixes and convert the NAVADMIN text to a
'*.ctxt' (canonical text) file if needed, and modify the web server to
serve the file if it's present, otherwise fallback to the source version
as needed.

I also make some manual reformats to NAVADMINs whose breakage is too
rare to warrant fixing the auto-script.