Add utility to try to decode NAVADMIN into header and msg body.
This isn't perfect but after some testing it seems to be able to handle
all messages in 2020 and onward, except that it still fails on messages
that are PERSONAL FOR or noting an unnumbered NAVADMIN, and these
messages actually *don't* have a body (but for now we still fail on
these).

This should give the ability to try to find the REF/ and AMPN/ fields in
the header which should eventually give the ability to cross-reference
NAVADMINs to pertinent policies and back.

But for now I just want to be able to split NAVADMINs into headers and
bodies before going deeper.

To validate the output of this script, I've been using a command like

    parallel -u --joblog jobs.tsv ./split-msg.pl ::: NAVADMIN/NAV20*.txt > /dev/null

to generate a table of all NAVADMINs for a given year (2020 in this
case), ignoring output for now.  In this table, the seventh column is
the exit status, the script gives exit status 1 if it fails to detect a
message body and you can use awk to print the offending file name
(column 10) for just those job runs when the command failed using a
command like:

    cat jobs.tsv | awk '$7==1 {print $10}' > failed-navadmins.txt