Add support for other hyperlinks in per-NAVADMIN metadata.
This permits adding hyperlinks in navadmin_meta.json in a "links"
object, which is an array of objects.  Each object contains a "url"
string and a "text" link.

If present, this object will cause the show-navadmin template to
generate additional button links just before the NAVADMIN is shown.

This can be useful to call attention to action items in the NAVADMIN,
however it requires manually pulling out hyperlinks into the metadata to
call attention to them.