Retry failed downloads with ?ver= string if needed.
Most NAVADMIN hyperlinks do not require the Sharepoint ?ver= add-on
query string parameter, so I normally download the link without it as
it is a cleaner link.

However some NAVADMINs, example NAVADMIN 216/23, will return a 404 error
from the server if the ?ver= query param is not included as specified in
the list of NAVADMINs.

This change still tries to download NAVADMINs without the ver add-on
first, but will retry the download with the original URL on failure.
This fixes NAVADMIN 216/23 at least, and hopefully future such issues.