Hyperlink known instructions page to appropriate cross-ref listing.
There remains a bug in which some instructions get a hyperlink but
loading the page shows an error that the page cannot be found.

For example, the listing of instructions shows a listing for OPNAVINST
1752.3B, but clicking on this link shows an error. Looking at the JSON
for the cross-references built by build-reference-list.pl, the entry for
OPNAVINST 1752.3B is present and points to NAVADMIN 002/17, but opening
that NAVADMIN shows that the reference is really for SECNAVINST 1752.3B
(and this hyperlink does work).

But, I believe this is due to an error in the generated list of
cross-references, either by the generator or internal to the web app
when it builds the overall list.