Commit 02b4b4c identified a bug in which some instructions that are hyperlinked in the Known Instructions page led to a broken link when clicked.
It turns out that this is due to the 3 instruction categories (DODINST,
SECNAVINST, OPNAVINST) all sharing the same data structure to hold all
instructions behind the scenes. This one single structure got assigned
instruction series for all 3 instruction categories, which meant that
every possible combination was generated on the page.

Changing "@foo{@keys} = ({}) x @keys" to "$foo{$_} = {} foreach @keys"
is sufficient to fix this issue, by creating a new empty hash ({}) for
every instruction.  Surprisingly (to me), the "({}) x @keys" operator
manages to alias the one {} for each repetition.

There remain bugs here, such as broken references to NAVADMINs in the
list for OPNAVINST 6110.1J, but this time it really *is* a data bug
instead of a bug in the web app, so that will be fixed separately.