Show something if NAVADMIN did not decode.
A NAVADMIN can fail to decode as a UTF-8 file with some source files
that have Windows-encoded stray characters. This is fixed when deploying
by running the miscoded files detector and fixer scripts, but this is
still annoying to have happen during local development.

Fix this by accepting the risk of stray weird characters during dev by
showing an escaped version of whatever loads if the file isn't in UTF-8
format.