Fix over-aggressive line splitting of NARR REF blocks.
Commit 8149ac4 noted that OPNAVINST 6110.1J cross-reference page
included many obviously-wrong listings, where the NAVADMIN name was
replaced with some kind of long NAVADMIN narrative line.

This turned out to be due to the NAVADMIN header decoder circuitry being
too aggressive in splitting run-on lines into separate lines in the
header.

This feature is needed to process older NAVADMINs, which had what had
previously been linebreaks stripped out in the version we have on disk.
Example from NAVADMIN 378/12:

    NARR/REF A IS TITLE 10, UNITED STATES CODE, SECTION 531// REF B IS OPNAVINST
    1210.5// REF C IS MILPERSMAN ARTICLE 1212-010// REF D IS MILPERSMAN ARTICLE
    1212-030//

The // is supposed to terminate each line (such that REF B would start
on its own line, REF C its own line, etc.).  The line splitter looks for
// in situations where it has additional characters after it but doesn't
end a line, and if it sees that adds the new lines back in.

When this is done to some NAVADMINs, it confuses a separate function
(the field detector) into thinking that a field is present.

For example (from NAVADMIN 276/17):

    NARR/REF A IS OPNAVINST 6110.1J, PHYSICAL READINESS PROGRAM.
    REF B IS OPNAVINST 1811.3A, VOLUNTARY RETIREMENT AND TRANSFER TO THE FLEET·
    RESERVE OF MEMBERS OF THE NAVY SERVING ON ACTIVE DUTY.
    REF C IS SECNAVINST 1920.6C, ADMINISTRATIVE SEPARATION OF OFFICERS.
    REF D IS MILPERSMAN ARTICLE 1070-180, OFFICER PHOTOGRAPHS.// REF E IS·
    NAVADMIN 178/15, PHYSICAL READINESS PROGRAM POLICY CHANGES.//

The last two lines get turned into *three*:

    REF D IS MILPERSMAN ARTICLE 1070-180, OFFICER PHOTOGRAPHS.//
    REF E IS·
    NAVADMIN 178/15, PHYSICAL READINESS PROGRAM POLICY CHANGES.//

REF E decodes to a null instruction and this completes handling of the
NARR field that had been under processing.

Subsequently, the field detector looks for the next field, finds
"NAVADMIN", and processes the line like any other field, overwriting the
already read-in value of the NAVADMIN field with gibberish on the rest
of the line.

The fix I chose was to be just a bit stricter, avoiding line splitting
if it looks like there was a REF field accidentally jammed together. You
can't go too aggressive with this check however (if you check for either
"REF/" or "REF " then you'll break older malformed NAVADMINs).

In testing, this update doesn't break any existing detection of cross
references, and in fact manages to find one more cross-reference (for
OPNAVINST 1210.5 which now correctly cross-references to NAVADMIN
378/12).