Note NAVADMINs which have a cross-reference but cannot be directly shown.
This can happen when the NAVADMIN is not stored on disk as a file named
'NAVddnnn.txt', such as when there are multiple files for the same
NAVADMIN, or stray letter suffixes.

This is most common on much older NAVADMINs. The newer ones typically
are silently replaced on the server instead.