Even more error-checking for NAVADMIN loading.
The Navy website now lists a directory of NAVADMINs (for year 2016) that
is now empty and returning a 404 instead.  Good thing I archived
those...

Fixed by moving the per-year download error check to only stop the
script if it's not refreshing a previously-downloaded prior year.

Other transient errors when debugging made clear that one of the error
messages is wrong. Fixed the error message.

Finally, added some slowdown to the part of the script that downloads
each individual years' worth of NAVADMIN subjects. This wasn't meant to
debug whether the server was itself dropping requests, but I was able to
confirm on a browser that the 2016 NAVADMINs are simply gone and it
wasn't an error with my script. But I'll leave it in.