Add updated (including very old) NAVADMINs.
The initial script run in the new year did a full scan of the source
website and that revealed a decent batch of new NAVADMINs from 1992-1995
relating to permissive TDY, and a 2004 NAVADMIN relating to the
Lautenberg gun control amendment.

Slight changes to the script were required to be able to display some of
these that have unusual filenames.